# apps/web/openapi.codegen.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "sahabatkreator API (codegen)",
    "version": "0.1.0"
  },
  "servers": [
    {
      "url": "http://localhost:3000",
      "description": "Local development server"
    }
  ],
  "paths": {
...
    "/v1/accounts": {
      "get": {
        "summary": "Daftar akun sosial",
        "description": "Semua akun sosial milik organisasi (termasuk status koneksi).",
        "tags": ["Accounts"],
        "x-required-scopes": ["accounts:read"],
        "security": [
          {
            "apiKeyAuth": []
          }
        ],
        "parameters": [],
        "responses": {
          "200": {
            "description": "Daftar akun.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AccountsResponse"
                }
              }
            }
          },
          "400": {
            "description": "Permintaan tidak valid (validasi schema / input salah).",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "401": {
            "description": "API key tidak ada, tidak valid, atau sudah dicabut.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "402": {
            "description": "Plan tidak termasuk api_access (atau api_write untuk endpoint tulis).",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "403": {
            "description": "Key tidak memiliki scope yang dibutuhkan endpoint ini.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "404": {
            "description": "Endpoint tidak ditemukan, atau sumber daya tidak ada.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "429": {
            "description": "Terlalu banyak permintaan (60/menit per key).",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        }
      }
    },
...
    "/v1/posts": {
      "get": {
        "summary": "Daftar post",
        "description": "Daftar grup post (draft, terjadwal, sudah tayang) milik organisasi.",
        "tags": ["Posts"],
        "x-required-scopes": ["posts:read"],
        "security": [
          {
            "apiKeyAuth": []
          }
        ],
        "parameters": [
          {
            "name": "from",
            "in": "query",
            "required": false,
            "schema": {
              "type": "string"
            },
            "description": "Filter YYYY-MM-DD"
          },
          {
            "name": "to",
            "in": "query",
            "required": false,
            "schema": {
              "type": "string"
            },
            "description": "Filter YYYY-MM-DD"
          },
          {
            "name": "status",
            "in": "query",
            "required": false,
            "schema": {
              "type": "string"
            },
            "description": "Filter status post"
          },
          {
            "name": "includeExternal",
            "in": "query",
            "required": false,
            "schema": {
              "type": "string",
              "enum": ["1"]
            },
            "description": "Sertakan post eksternal"
          },
          {
            "name": "page",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "default": 1
            }
          },
          {
            "name": "perPage",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "default": 50,
              "minimum": 1,
              "maximum": 100
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Daftar post.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PostsListResponse"
                }
              }
            }
          },
          "400": {
            "description": "Permintaan tidak valid (validasi schema / input salah).",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "401": {
            "description": "API key tidak ada, tidak valid, atau sudah dicabut.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "402": {
            "description": "Plan tidak termasuk api_access (atau api_write untuk endpoint tulis).",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "403": {
            "description": "Key tidak memiliki scope yang dibutuhkan endpoint ini.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "404": {
            "description": "Endpoint tidak ditemukan, atau sumber daya tidak ada.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "429": {
            "description": "Terlalu banyak permintaan (60/menit per key).",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        }
      },
      "post": {
        "summary": "Buat post",
        "description": "Membuat grup post baru. Bila `scheduledAt` diisi, menjadwalkannya (butuh fitur scheduled_posts).",
        "tags": ["Posts"],
        "x-required-scopes": ["posts:write"],
        "security": [
          {
            "apiKeyAuth": []
          }
        ],
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreatePostRequest"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Grup post dibuat.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CreatePostResponse"
                }
              }
            }
          },
          "400": {
            "description": "Permintaan tidak valid (validasi schema / input salah).",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "401": {
            "description": "API key tidak ada, tidak valid, atau sudah dicabut.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "402": {
            "description": "Plan tidak termasuk api_access (atau api_write untuk endpoint tulis).",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "403": {
            "description": "Key tidak memiliki scope yang dibutuhkan endpoint ini.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "404": {
            "description": "Endpoint tidak ditemukan, atau sumber daya tidak ada.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "429": {
            "description": "Terlalu banyak permintaan (60/menit per key).",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        }
      }
    },
...
    "/v1/media": {
      "get": {
        "summary": "Daftar media",
        "description": "Item media library (maks 200).",
        "tags": ["Media"],
        "x-required-scopes": ["media:read"],
        "security": [
          {
            "apiKeyAuth": []
          }
        ],
        "parameters": [
          {
            "name": "folderId",
            "in": "query",
            "required": false,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Daftar media.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/MediaListResponse"
                }
              }
            }
          },
          "400": {
            "description": "Permintaan tidak valid (validasi schema / input salah).",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "401": {
            "description": "API key tidak ada, tidak valid, atau sudah dicabut.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "402": {
            "description": "Plan tidak termasuk api_access (atau api_write untuk endpoint tulis).",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "403": {
            "description": "Key tidak memiliki scope yang dibutuhkan endpoint ini.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "404": {
            "description": "Endpoint tidak ditemukan, atau sumber daya tidak ada.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "429": {
            "description": "Terlalu banyak permintaan (60/menit per key).",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        }
      }
    },
    "/v1/media/import": {
      "post": {
        "summary": "Import media dari URL",
        "description": "Mengunduh file dari URL publik ke media library organisasi.",
        "tags": ["Media"],
        "x-required-scopes": ["media:write"],
        "security": [
          {
            "apiKeyAuth": []
          }
        ],
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ImportMediaRequest"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Media diimpor.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ImportMediaResponse"
                }
              }
            }
          },
          "400": {
            "description": "Permintaan tidak valid (validasi schema / input salah).",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "401": {
            "description": "API key tidak ada, tidak valid, atau sudah dicabut.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "402": {
            "description": "Plan tidak termasuk api_access (atau api_write untuk endpoint tulis).",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "403": {
            "description": "Key tidak memiliki scope yang dibutuhkan endpoint ini.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "404": {
            "description": "Endpoint tidak ditemukan, atau sumber daya tidak ada.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "429": {
            "description": "Terlalu banyak permintaan (60/menit per key).",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "503": {
            "description": "Storage R2 belum dikonfigurasi.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "ErrorResponse": {
        "type": "object",
        "properties": {
          "message": {
            "type": "string"
          }
        },
        "required": ["message"]
      },
...
      "AccountsResponse": {
        "type": "object",
        "properties": {
          "accounts": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/AccountItem"
            }
          }
        },
        "required": ["accounts"]
      },
...
      "PostsListResponse": {
        "type": "object",
        "properties": {
          "groups": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/PostGroupItem"
            }
          },
          "page": {
            "type": "number"
          },
          "perPage": {
            "type": "number"
          }
        },
        "required": ["groups", "page", "perPage"]
      },
...
      "CreatePostRequest": {
        "type": "object",
        "properties": {
          "content": {
            "type": "string",
            "default": ""
          },
          "scheduledAt": {
            "type": ["string", "null"],
            "format": "date-time"
          },
          "timezone": {
            "type": "string",
            "default": "Asia/Jakarta"
          },
          "audioTrackId": {
            "type": ["string", "null"]
          },
          "productIds": {
            "type": "array",
            "items": {
              "type": "string"
            },
            "default": []
          },
          "items": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/CreatePostItem"
            },
            "minItems": 1
          }
        },
        "required": ["items"]
      },
...
      "CreatePostResponse": {
        "type": "object",
        "properties": {
          "postGroupId": {
            "type": "string"
          }
        },
        "required": ["postGroupId"]
      },
...
      "MediaListResponse": {
        "type": "object",
        "properties": {
          "items": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/MediaItem"
            }
          },
          "storageConfigured": {
            "type": "boolean"
          }
        },
        "required": ["items", "storageConfigured"]
      },
      "ImportMediaRequest": {
        "type": "object",
        "properties": {
          "url": {
            "type": "string",
            "maxLength": 2048,
            "format": "uri"
          },
          "name": {
            "type": "string",
            "maxLength": 200
          },
          "folderId": {
            "type": ["string", "null"]
          }
        },
        "required": ["url"]
      },
      "ImportMediaResponse": {
        "type": "object",
        "properties": {
          "media": {
            "$ref": "#/components/schemas/MediaItem"
          }
        },
        "required": ["media"]
      },
...
    },
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT"
      },
      "apiKeyAuth": {
        "type": "http",
        "scheme": "bearer",
        "description": "API key format `sk_api_<32 karakter>`. Dibuat di Settings → API. Bisa juga dikirim via header `X-API-Key`."
      }
    }
  }
}